+++ numbers/in_range.go
@@ -6,35 +6,31 @@ import (
 	"github.com/golodash/godash/internal"
 )
 
-// Check number lower than end and greater then start
+// Check if number is lower than 'upper' and greater than 'lower'.
 //
 // Complexity: O(1)
-func InRange(number, start, end interface{}) interface{} {
+func InRange(number, lower, upper interface{}) interface{} {
 	if !internal.IsNumber(number) {
 		panic("'number' is not a number")
 	}
-	if !internal.IsNumber(start) {
+	if !internal.IsNumber(lower) {
 		panic("'start' is not a number")
 	}
-	if !internal.IsNumber(end) {
+	if !internal.IsNumber(upper) {
 		panic("'end' is not a number")
 	}
 
-	numberValue := reflect.ValueOf(number).Convert(reflect.TypeOf(1.0)).Float()
-	startValue := reflect.ValueOf(start).Convert(reflect.TypeOf(1.0)).Float()
-	endValue := reflect.ValueOf(end).Convert(reflect.TypeOf(1.0)).Float()
+	numberFloat := reflect.ValueOf(number).Convert(reflect.TypeOf(1.0)).Float()
+	lowerFloat := reflect.ValueOf(lower).Convert(reflect.TypeOf(1.0)).Float()
+	upperFloat := reflect.ValueOf(upper).Convert(reflect.TypeOf(1.0)).Float()
 
-	if startValue <= endValue {
-		if numberValue >= startValue && numberValue < endValue {
-			return true
-		} else {
-			return false
-		}
+	if lowerFloat > upperFloat {
+		panic("'upper' has to be higher or equal higher than 'lower'")
+	}
+
+	if numberFloat >= lowerFloat && numberFloat <= upperFloat {
+		return true
 	} else {
-		if numberValue >= endValue && numberValue < startValue {
-			return true
-		} else {
-			return false
-		}
+		return false
 	}
 }